--- src/tasks/tasks.hpp
@@ -1,3 +1,0 @@
-void connectWifi (void *pvParameters);
-void connectThingsBoard (void *pvParameters);
-void sendTelemetryData (void *pvParameters);
